+++ graphene_django_crud/input_types.py
@@ -10,13 +10,16 @@ def gen_inputType_filter(
     with_text_filter=False,
     with_date_filter=False,
     with_time_filter=False,
+    with_isnull_filter=True,
 ):
 
     items = OrderedDict()
 
     items["equals"] = graphene.Field(field)
     items["in"] = graphene.List(field)
-    items["isnull"] = graphene.Boolean()
+
+    if with_isnull_filter:
+        items["isnull"] = graphene.Boolean()
 
     if with_number_filter:
         items["gt"] = graphene.Field(field)
@@ -47,6 +50,8 @@ def gen_inputType_filter(
         items,
     )
 
+
+IdFilter = gen_inputType_filter("IdFilter", graphene.ID, with_isnull_filter=False)
 
 IntFilter = gen_inputType_filter(
     "IntFilter", graphene.Int, with_number_filter=True, with_text_filter=True